pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "devgiordane",email = "bruno_giordane@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=4.0.0",
    "dash-bootstrap-components>=1.6.0",
    "dash-ag-grid>=31.0.0",
    "sqlalchemy>=2.0.0",
    "plotly>=5.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "flask>=3.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
